Hack to enable '=' in the middle of command arguments.
The problem is that '=' cannot be used in command arguments
qithout quoting to avoid the ambiguity below:

  ls = "a"

It's a variable assignment? or a 'ls' listing on files '=' and "a"?

This commit leaves the problem above as is, but enable '=' in other
cases, like below:

  make collection=project1

or

  docker run -e VAR1= -e VAR2= something

But EQUAL is still *prohibited* in the cases below:

  make otheroption =something
  make otheroption = othervalue
  make = something # this is an invalid assignment due to missing quote

Maybe this confuses the syntax, and people could find this behaviour
more *surprise* than the enforcement to quote always...