Beginning of a bigger refactor in the interpreter
Now instead of treat a function as a whole interpreter
we have a single interpreter using lambda functions
as environment (storages). The thing is incomplete for
now but this will fix our closure bugs because now
a function declaration when evaluated will generate a
FnDef (function definition) that when executed must
instantiate a new lambda passing the closure (if any)
explicitly, then there's no sharing between function
calls anymore.

Signed-off-by: Tiago Natel de Moura <tiago.natel@neoway.com.br>