Adds the concept of function definition (sh.FnDef)
Now when interpreter parses ast.FnDecl it generates a
function definition (sh.FnDef) that stores the meta
data required to build a function. In case of a
function invocation (eval of ast.FnInv) the interpreter
now need to first build the function from the FnDef.
This guarantees a new fresh environment at every call.

Signed-off-by: Tiago Natel de Moura <tiago.natel@neoway.com.br>