OSX symlinks of /tmp and /var break the tests
On OSX, the directories /tmp and /var are symlinks to
/private/tmp and /private/var respectively.
This change adds an additional call to filepath.EvalSymlinks
before using paths obtained by TempDir.

Also, this change identified a problem related to logging of
the warnings when nashpath and/or nashroot aren't properly tested.
As there are tests that do instantiate nash with wrong directories
for testing, this was making a mess in the stderr of the tests
because now nash doesn't abort in case of errors to easy the install
process. This change address that too.

Signed-off-by: Tiago de Bem <tiago.moura@nginx.com>