Just create offset var when it is defined
The idea is to make it more explicit that on error situations
the offset returned is zero. The function was doing that but
in a way that we had to check if the offset var was
not being assigned to some value, in the end it never is,
it is always 0 (or I got something wrong).